refactor(step4): 테스트코드 경로맞추기
패키지 경로 맞추기
