+++ src/test/java/ladder/domain/DirectionTest.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 import ladder.domain.Direction;
 import org.junit.jupiter.api.DisplayName;
