+++ src/test/java/ladder/domain/PlayerTest.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 import ladder.domain.Player;
 import org.junit.jupiter.api.DisplayName;
