+++ src/test/java/ladder/domain/PrizesTest.java
@@ -1,18 +1,20 @@
-package ladder;
+package ladder.domain;
 
-import ladder.domain.Prize;
+import ladder.domain.Prizes;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import java.util.Arrays;
+
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-public class PrizeTest {
+public class PrizesTest {
 
     @Test
-    @DisplayName("게임결과 생성 실패 테스트")
+    @DisplayName("플레이결과들 생성 실패 테스트")
     void validateTest() {
         assertThatThrownBy(
-                () -> Prize.of("팔굽혀펴기40회")
+                () -> new Prizes(Arrays.asList("asdf"))
         ).isInstanceOf(IllegalArgumentException.class);
     }
 }
